commons/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile commons/requirements.in
#
alembic==1.8.1
    # via -r commons/requirements.in
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
boto3==1.24.59
    # via -r commons/requirements.in
botocore==1.27.59
    # via
    #   boto3
    #   s3transfer
cachetools==5.2.0
    # via -r commons/requirements.in
celery[redis]==5.2.7
    # via -r commons/requirements.in
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
deprecated==1.2.13
    # via redis
dnspython==2.2.1
    # via email-validator
ecdsa==0.18.0
    # via python-jose
email-validator==1.2.1
    # via -r commons/requirements.in
fastapi==0.80.0
    # via -r commons/requirements.in
greenlet==1.1.3
    # via sqlalchemy
gunicorn==20.1.0
    # via -r commons/requirements.in
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via mako
orjson==3.7.12
    # via -r commons/requirements.in
packaging==21.3
    # via redis
prompt-toolkit==3.0.30
    # via click-repl
psycopg2==2.9.3
    # via -r commons/requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==1.9.2
    # via
    #   -r commons/requirements.in
    #   fastapi
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via botocore
python-dotenv==0.20.0
    # via uvicorn
python-jose==3.3.0
    # via -r commons/requirements.in
pytz==2022.2.1
    # via celery
pyyaml==6.0
    # via
    #   -r commons/requirements.in
    #   uvicorn
redis==4.3.4
    # via
    #   -r commons/requirements.in
    #   celery
requests==2.28.1
    # via -r commons/requirements.in
rsa==4.9
    # via python-jose
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   ecdsa
    #   python-dateutil
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.40
    # via
    #   -r commons/requirements.in
    #   alembic
    #   sqlalchemy-utils
sqlalchemy-utils==0.38.3
    # via -r commons/requirements.in
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.12
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.18.3
    # via -r commons/requirements.in
uvloop==0.16.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.16.1
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.3
    # via uvicorn
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
